+++ CHANGELOG.md
@@ -17,6 +17,14 @@ Example: 10 mar 21:13 SAJID\
 * dbase folder --> added files for patient appointments, and doctor availability
 
 
+## 27 MARCH Jenny
+* Removed availability field in DoctorModel. 
+
+## 26 MARCH Jenny
+# 10:46 
+* Rehaul of Schedule and how appointments are stored/view with Sajid. Commenting out all information in Doctor MVC related to Schedule.
+# 10:57
+* Adding new field to DoctorModel constructor.
 
 ## 25 MARCH 21:44 Jenny
 Overview/Summary: